requirements.txt
aerich==0.8.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8
asyncpg==0.30.0
atlastk==0.13.4
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dictdiffer==0.9.0
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iso8601==2.1.0
jinja2==3.1.6
jwt==1.3.1
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
passlib==1.7.4
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
pypika-tortoise==0.5.0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyyaml==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.1
tortoise==0.1.1
tortoise-orm==0.24.2
typer==0.15.2
typing-extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
